[Windows] Fix Visual Studio warnings (we have quite a lot of them)

+++ xbmc/cores/VideoPlayer/VideoPlayerAudio.cpp
@@ -342,7 +342,7 @@ void CVideoPlayerAudio::Process()
       {
         m_dvdAudio.Pause();
       }
-      m_speed = speed;
+      m_speed = (int)speed;
     }
     else if (pMsg->IsType(CDVDMsg::GENERAL_STREAMCHANGE))
     {
